twin/consumer/requirements.txt
kafka-python
avro-python3